Add final dao unit tests [#21]

+++ src/main/java/com/teamlans/lepta/database/daos/BillDaoImpl.java
@@ -1,8 +1,6 @@
 package com.teamlans.lepta.database.daos;
 
 import com.teamlans.lepta.database.entities.Bill;
-import com.teamlans.lepta.database.entities.Item;
-import com.teamlans.lepta.database.enums.Status;
 import com.teamlans.lepta.service.bill.BillService;
 
 import org.hibernate.Session;
@@ -10,13 +8,10 @@ import org.hibernate.SessionFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
+import org.springframework.dao.DataRetrievalFailureException;
 import org.springframework.stereotype.Repository;
 
-import java.util.Date;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 @Repository
 public class BillDaoImpl implements BillDao {
@@ -37,6 +32,9 @@ public class BillDaoImpl implements BillDao {
     logger.debug("Deleting bill with id {}", nr);
     Session session = factory.getCurrentSession();
     Bill bill = session.get(Bill.class, nr);
+    if (bill == null) {
+      throw new DataRetrievalFailureException("Could not delete bill with id " + nr);
+    }
     session.delete(bill);
   }
 
@@ -51,31 +49,7 @@ public class BillDaoImpl implements BillDao {
   public void updateBill(Bill newBill) {
     logger.debug("Updating bill with {}", newBill);
     Session session = factory.getCurrentSession();
-
-    Integer nr = newBill.getNr();
-    Bill bill = session.get(Bill.class, nr);
-
-    Status newStatus = newBill.getStatus();
-    bill.setStatus(newStatus);
-
-    Date newTimestamp = newBill.getDate();
-    bill.setDate(newTimestamp);
-
-    Set<Item> newItems = newBill.getItems();
-    Set<Item> items = bill.getItems();
-    Set<Item> deletedItems = new HashSet<>();
-    for (Item item : items) {
-      if (!newItems.contains(item)) {
-        deletedItems.add(item);
-      }
-    }
-    bill.removeItems(deletedItems);
-    for (Item newItem : newItems) {
-      if (!items.contains(newItem)) {
-        bill.addItem(newItem);
-      }
-    }
-    session.update(bill);
+    session.update(newBill);
   }
 }
